Fix OverlayArrow being misaligned when offscreen
The coordinates of the target widget were being calculated from the
arrow widget, which is drawn inside a leaflet page that slides across
the screen. To solve this problem, we calculate the coordinates based on
the leaflet page's parent widget.

This is a bit crude, but there is a reasonable path from this state to
something more broadly useful.